+++ cookeat.zip_expanded/cookeat/src/main/java/com/horizon/cookeat/entities/RecipeIngredient.java
@@ -42,7 +42,7 @@ public class RecipeIngredient implements Serializable{
     private Ingredient ingredient;
  
     @Column(name = "quantity")
-    private int quantity;
+    private float quantity;
  
     // METHODS //
     
@@ -65,7 +65,7 @@ public class RecipeIngredient implements Serializable{
     // CONSTRUCTOR //
     public RecipeIngredient() {}
     
-    public RecipeIngredient(Recipe r, Ingredient i, int quantity) {
+    public RecipeIngredient(Recipe r, Ingredient i, float quantity) {
         this.recipe = r;
         this.ingredient = i;
 
@@ -100,11 +100,11 @@ public class RecipeIngredient implements Serializable{
 		this.ingredient = ingredient;
 	}
 
-	public int getQuantity() {
+	public float getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(int quantity) {
+	public void setQuantity(float quantity) {
 		this.quantity = quantity;
 	}
